+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+**Features**:
+
+- Scrub non-minidump attachments if there are explicit `$attachment` rules. ([#4415](https://github.com/getsentry/relay/pull/4415))
+
 **Internal**:
 
 - Remove the `spool` command from Relay. ([#4423](https://github.com/getsentry/relay/pull/4423))